feat: UUID-only relay connection — remove attach token (founder decision)
The routing UUID in the relay wss URL is now the sole bearer credential
for agent-to-relay connections. Removes the optional attach-token/secret
(--remote-secret / VIBE_REMOTE_SECRET) from the MCP CLI, browser CLI, and
set_remote MCP tool. Agents connect with only --remote <uuid-or-url>; if
the UUID leaks, regenerate it in the Vibe extension Settings.

Replaces scripts/e2e-remote-auth.mjs (obsolete bearer-token test) with
scripts/e2e-uuid-only-auth.mjs, which asserts no Authorization header is
ever sent, --remote-secret is now a rejected unknown option, and a stray
VIBE_REMOTE_SECRET env var has no effect. Updates README.md, packages/cli/README.md,
docs/openclaw-local-browser.md, docs/chrome-devtools-relay.md, and
openclaw/vibebrowser/SKILL.md to drop the two-value auth model.

Composes with the parallel relay-server + extension change in the main
vibe repo (feat/relay-uuid-only-auth).
