+++ package.json
@@ -23,13 +23,13 @@
     "start": "node dist/cli.js",
     "validate:skill": "node scripts/validate-skill.mjs openclaw/vibebrowser/SKILL.md",
     "prepublishOnly": "npm run build",
-    "test": "npm run test:e2e:relay-race && npm run test:e2e:relay-roundtrip && npm run test:e2e:cli-relay && npm run test:e2e:http && npm run test:e2e:remote-auth && npm run test:e2e:tools-list-budget && npm run test:e2e:browser-cli && npm run test:e2e:cli-package && npm run test:e2e:devtools-flag",
-    "test:ci": "npm run test:e2e:relay-race && npm run test:e2e:relay-roundtrip && npm run test:e2e:cli-relay && npm run test:e2e:http && npm run test:e2e:remote-auth && npm run test:e2e:tools-list-budget && npm run test:e2e:browser-cli && npm run test:e2e:cli-package && npm run test:e2e:devtools-flag",
+    "test": "npm run test:e2e:relay-race && npm run test:e2e:relay-roundtrip && npm run test:e2e:cli-relay && npm run test:e2e:http && npm run test:e2e:uuid-only-auth && npm run test:e2e:tools-list-budget && npm run test:e2e:browser-cli && npm run test:e2e:cli-package && npm run test:e2e:devtools-flag",
+    "test:ci": "npm run test:e2e:relay-race && npm run test:e2e:relay-roundtrip && npm run test:e2e:cli-relay && npm run test:e2e:http && npm run test:e2e:uuid-only-auth && npm run test:e2e:tools-list-budget && npm run test:e2e:browser-cli && npm run test:e2e:cli-package && npm run test:e2e:devtools-flag",
     "test:e2e:tools-list-budget": "node scripts/e2e-tools-list-startup-budget.mjs",
     "test:e2e:browser-cli": "node scripts/e2e-browser-cli.mjs",
     "test:e2e:cli-relay": "node scripts/e2e-cli-relay.mjs",
     "test:e2e:http": "node scripts/e2e-http-streamable.mjs",
-    "test:e2e:remote-auth": "node scripts/e2e-remote-auth.mjs",
+    "test:e2e:uuid-only-auth": "node scripts/e2e-uuid-only-auth.mjs",
     "test:e2e:devtools-flag": "node scripts/e2e-devtools-flag.mjs",
     "test:e2e:cli-package": "node scripts/e2e-cli-package-smoke.mjs",
     "test:e2e:browser-cli-live": "node scripts/e2e-browser-cli-live.mjs",
